+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Hughes 2024 vancomycin parametric and nonparametric models ([doi:10.1097/FTD.0000000000001214](https://doi.org/10.1097/FTD.0000000000001214)) -- hospitalized adults with class 3 obesity.
+
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.
 
 - Add Ravva 2010 varenicline exposure-response models ([doi:10.1038/clpt.2009.282](https://doi.org/10.1038/clpt.2009.282)) -- adult cigarette smokers in smoking-cessation trials.